requirements.txt
# Core LangChain & LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5

# DeepSeek (compatible with OpenAI API)
openai>=1.12.0

# Streamlit UI
streamlit>=1.31.0

# Environment & Config
python-dotenv>=1.0.0

# MCP SDK
mcp>=0.9.0

# GitHub API
PyGithub>=2.1.1

# Google Drive API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.2.0

# HTTP requests
requests>=2.31.0

# Machine Learning
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
kaggle>=1.6.0
huggingface-hub>=0.20.0
joblib>=1.3.0

# Utilities
pydantic>=2.5.0
typing-extensions>=4.9.0


